Refactor CertificateCard to Integrate Certificate Processing Hook
- Introduce useCertificateProcessor hook to streamline certificate generation logic
- Replace the previous handleGenerate and generateCertificate functions with a more organized approach using the new hook
- Enhance code readability and maintainability by consolidating certificate processing logic
